pyproject.toml
[project]
name = "weather-api-investigation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "folium>=0.20.0",
    "meteostat>=1.7.4",
    "numpy>=2.0.2",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
]
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"
